Updated remote_setup_pc to connect to the Internet then Movo.
This change enables the use case of a researcher wanting to set up a computer for simulation and future connection to the Movo without physical access to the robot during setup. The existing structure of the script requires the computer to be connected to both the Movo and the Internet from the start, even though the contents of the script do one then the other.